requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.6
django-import-export==2.5.0
django-pandas==0.6.4
et-xmlfile==1.1.0
gunicorn==20.1.0
joblib==1.0.1
MarkupPy==1.14
mecab-python3==1.0.4
numpy==1.21.1
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.3.2
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
threadpoolctl==2.2.0
xlrd==2.0.1
xlwt==1.3.0
